# polynomials: create coefficients in cbd from an iterator